% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_truth.R
\name{compare_truth}
\alias{compare_truth}
\title{Compare simulation estimates to truth}
\usage{
compare_truth(a1, trt1 = NA, a2 = NA, trt2 = NA, effect, marginal,
  list_of_simulation_estimates, truth)
}
\description{
Compare simulation estimates to truth
}

