% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_truth.R
\name{summarize_comparisons}
\alias{summarize_comparisons}
\title{Summarize comparisions to truth}
\usage{
summarize_comparisons(estimates, truth, alphas)
}
\description{
Summarize comparisions to truth
}

