% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimands.R
\name{ybar_i_aalpha}
\alias{ybar_i_aalpha}
\title{Individual level causal estimand Y(a, alpha)}
\usage{
ybar_i_aalpha(outcomes, alpha)
}
\value{
n x 2 matrix with values of ybar_ij(a, alpha)
}
\description{
Takes n X n  x 2 array as input
}

